+++ src/main/java/model/Lotto.java
@@ -1,6 +1,7 @@
 package model;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
@@ -10,6 +11,7 @@ import static java.util.stream.Collectors.toList;
 public class Lotto {
     private static final int LOTTO_SIZE = 6;
     private static final String LOTTO_SIZE_MUST_BE_SIX = "로또 번호는 중복되지 않는 6개의 번호여야 합니다.";
+    private static final String SEPARATOR = ",";
 
     private final List<LottoNumber> numbers;
 
@@ -41,6 +43,18 @@ public class Lotto {
 
     public boolean hasNumber(LottoNumber lottoNumber) {
         return numbers.contains(lottoNumber);
+    }
+
+    public static Lotto generateLottoByInput(String inputLottoNumbers) {
+        List<LottoNumber> lottoNumbers = generateLottoNumbersByInput(inputLottoNumbers);
+        return new Lotto(lottoNumbers);
+    }
+
+    private static List<LottoNumber> generateLottoNumbersByInput(String inputWinningLottoNumbers) {
+        return Arrays.stream(inputWinningLottoNumbers.split(SEPARATOR))
+                .map(Integer::valueOf)
+                .map(LottoNumber::new)
+                .collect(toList());
     }
 
     @Override
