+++ src/main/java/model/Lottos.java
@@ -3,6 +3,7 @@ package model;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.*;
+import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.*;
 
@@ -20,6 +21,32 @@ public class Lottos {
 
     public List<Lotto> getLottos() {
         return new ArrayList<>(lottos);
+    }
+
+    public static Lottos issueLottos(Money money) {
+        int issueCount = calculateIssueCount(money);
+        List<Lotto> lottos = generateLottos(issueCount);
+        return new Lottos(lottos);
+    }
+
+    private static List<Lotto> generateLottos(int issueCount) {
+        return Stream.generate(Lottos::generateLotto)
+                .limit(issueCount)
+                .collect(toList());
+    }
+
+    private static Lotto generateLotto() {
+        try {
+            return new Lotto(Lotto.generateLottoNumbers());
+        } catch (IllegalArgumentException e) {
+            return generateLotto();
+        }
+    }
+
+    private static int calculateIssueCount(Money money) {
+        return money.getAmount()
+                .divide(new BigDecimal(LOTTO_PRICE))
+                .intValue();
     }
 
     public Map<Rank, Integer> getStatistics(WinningLotto winningLotto) {
