+++ src/main/java/model/WinningLotto.java
@@ -30,6 +30,10 @@ public class WinningLotto {
                 .count();
     }
 
+    public static WinningLotto issueWinningLotto(Lotto lotto, LottoNumber bonusNumber) {
+        return new WinningLotto(lotto, bonusNumber);
+    }
+
     public Lotto getLotto() {
         return lotto;
     }
